Move RDF rules to their own folder (#fix #79)
Related to #78; this moves the RDF rules into their own folder.

Next step would be to move rule tests into their own sections.
