+++ tests/helpers.js
@@ -1,12 +1,10 @@
 import { singleLineString } from 'utils';
 
-
 export const fakeMessageData = {
   code: 'WHATEVER_CODE',
   description: 'description',
   message: 'message',
 };
-
 
 export function validHTML(contents='') {
   return singleLineString`<!DOCTYPE html>
@@ -19,4 +17,14 @@ export function validHTML(contents='') {
       ${contents}
     </body>
   </html>`;
+}
+
+export function validRDF(contents) {
+  return singleLineString`<?xml version='1.0' encoding='utf-8'?>
+  <RDF xmlns="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
+       xmlns:em="http://www.mozilla.org/2004/em-rdf#">
+    <Description about="urn:mozilla:install-manifest">
+      ${contents}
+    </Description>
+  </RDF>`;
 }
