+++ tests/validators/test.rdf.js
@@ -1,21 +1,12 @@
 import fs from 'fs';
 
-import { VALIDATION_ERROR, VALIDATION_NOTICE, VALIDATION_WARNING } from 'const';
+import sinon from 'sinon';
+import XMLDom from 'xmldom';
+
 import { RDFParseError } from 'exceptions';
+import { validRDF } from '../helpers';
 import RDFScanner from 'validators/rdf';
-import * as messages from 'messages';
-import { singleLineString } from 'utils';
 
-
-function validRDF(contents) {
-  return singleLineString`<?xml version='1.0' encoding='utf-8'?>
-  <RDF xmlns="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-       xmlns:em="http://www.mozilla.org/2004/em-rdf#">
-    <Description about="urn:mozilla:install-manifest">
-      ${contents}
-    </Description>
-  </RDF>`;
-}
 
 describe('RDF', function() {
 
@@ -39,89 +30,49 @@ describe('RDF', function() {
       });
   });
 
-  it('should not allow <hidden> tag', () => {
-    var contents = validRDF('<em:hidden>true</em:hidden>');
+  it('should run all rules in rules/html', () => {
+    var contents = validRDF();
     var rdfScanner = new RDFScanner(contents, 'install.rdf');
+    var fakeRules = {
+      iAmAFakeRule: sinon.stub(),
+      iAmAAnotherFakeRule: sinon.stub(),
+    };
 
-    return rdfScanner.mustNotExist()
+    return rdfScanner.scan(fakeRules)
       .then(() => {
-        var validatorMessages = rdfScanner.validatorMessages;
-
-        assert.equal(validatorMessages.length, 1);
-        assert.equal(validatorMessages[0].code,
-                     messages.TAG_NOT_ALLOWED_HIDDEN.code);
-        assert.equal(validatorMessages[0].severity, VALIDATION_ERROR);
+        assert.ok(fakeRules.iAmAFakeRule.calledOnce);
+        assert.ok(fakeRules.iAmAAnotherFakeRule.calledOnce);
       });
   });
 
-  it('should not blow up when multiple bad tags are found', () => {
-    var contents = validRDF(singleLineString`<em:hidden>true</em:hidden>
-      <em:hidden>false</em:hidden>`);
+  it('should return an already-parsed xmlDoc if exists', () => {
+    sinon.spy(XMLDom, 'DOMParser');
+
+    var contents = validRDF();
     var rdfScanner = new RDFScanner(contents, 'install.rdf');
 
-    return rdfScanner.mustNotExist()
+    return rdfScanner.getXMLDoc()
       .then(() => {
-        var validatorMessages = rdfScanner.validatorMessages;
-
-        assert.equal(validatorMessages.length, 2);
-        for (let message of validatorMessages) {
-          assert.equal(message.code, messages.TAG_NOT_ALLOWED_HIDDEN.code);
-          assert.equal(message.severity, VALIDATION_ERROR);
-        }
-      });
-  });
-
-  it('should not allow certain tags contextually (eg. when listed)', () => {
-    // Should fail because Add-on is listed and has an updateURL.
-    var contents = validRDF(singleLineString`<em:listed>true</em:listed>
-      <em:updateURL>http://mozilla.com/updateMyAddon.php</em:updateURL>`);
-    var rdfScanner = new RDFScanner(contents, 'install.rdf');
-
-    return rdfScanner.mustNotExist()
-      .then(() => {
-        var validatorMessages = rdfScanner.validatorMessages;
-
-        assert.equal(validatorMessages.length, 1);
-        for (let message of validatorMessages) {
-          assert.equal(message.code, messages.TAG_NOT_ALLOWED_UPDATEURL.code);
-          assert.equal(message.severity, VALIDATION_ERROR);
-        }
-
-        // This shouldn't fail because there is no listed tag.
-        contents = validRDF(singleLineString`
-          <em:updateURL>http://mozilla.com/updateMyAddon.php</em:updateURL>`);
-        rdfScanner = new RDFScanner(contents, 'install.rdf');
-
-        return rdfScanner.mustNotExist();
+        return rdfScanner.getXMLDoc();
       })
       .then(() => {
-        var validatorMessages = rdfScanner.validatorMessages;
-
-        assert.equal(validatorMessages.length, 0);
+        assert.ok(XMLDom.DOMParser.calledOnce);
+        XMLDom.DOMParser.restore();
       });
   });
 
-  it('should test for obsolete tags', () => {
-    var contents = validRDF(singleLineString`<em:file>'foo.js'</em:file>
-      <em:requires>'something'</em:requires><em:skin>true</em:skin>`);
+  it('should run all rules in rules/rdf', () => {
+    var contents = validRDF();
     var rdfScanner = new RDFScanner(contents, 'install.rdf');
+    var fakeRules = {
+      iAmAFakeRule: sinon.stub(),
+      iAmAAnotherFakeRule: sinon.stub(),
+    };
 
-    return rdfScanner.mustNotExist()
+    return rdfScanner.scan(fakeRules)
       .then(() => {
-        var validatorMessages = rdfScanner.validatorMessages;
-
-        assert.equal(validatorMessages.length, 3);
-
-        for (let message of validatorMessages) {
-          assert.equal(message.severity, VALIDATION_WARNING);
-        }
-
-        assert.equal(validatorMessages[0].code,
-                     messages.TAG_OBSOLETE_FILE.code);
-        assert.equal(validatorMessages[1].code,
-                     messages.TAG_OBSOLETE_REQUIRES.code);
-        assert.equal(validatorMessages[2].code,
-                     messages.TAG_OBSOLETE_SKIN.code);
+        assert.ok(fakeRules.iAmAFakeRule.calledOnce);
+        assert.ok(fakeRules.iAmAAnotherFakeRule.calledOnce);
       });
   });
 
@@ -138,20 +89,4 @@ describe('RDF', function() {
       });
   });
 
-  it('_checkForTags should find tags and set the correct severity', () => {
-    var contents = validRDF(singleLineString`<em:nameTag>Matthew Riley
-      MacPherson</em:nameTag><em:file>'foo.js'</em:file>`);
-    var rdfScanner = new RDFScanner(contents, 'install.rdf');
-
-    return rdfScanner
-      ._checkForTags(['file'], VALIDATION_NOTICE, 'TAG_OBSOLETE_')
-      .then(() => {
-        var validatorMessages = rdfScanner.validatorMessages;
-
-        assert.equal(validatorMessages.length, 1);
-        assert.equal(validatorMessages[0].code,
-                     messages.TAG_OBSOLETE_FILE.code);
-        assert.equal(validatorMessages[0].severity, VALIDATION_NOTICE);
-      });
-  });
 });
